+++ hyperbolic/src/solver.h
@@ -9,7 +9,7 @@ thomas(int n, float *a, float *b, float *c, float *d_x,
        float *d_z, float *u_x, float *u_z);
 
 int
-thomas_block(int n, double *a, double *b, double *c, double *d,
-             double *xz, double *D, double *y);
+thomas_block(int n, float *a, float *b, float *c, float *d,
+             float *xz, float *D, float *y);
 
 #endif
