+++ parabolic/lib/lib_math.c
@@ -4,11 +4,11 @@
 #include "lib_math.h"
 
 int
-mat_invert2x2(double matrix[][2])
+mat_invert2x2(float matrix[][2])
 {
   for (int k=0; k<2; k++)
   {
-     double con = matrix[k][k];
+     float con = matrix[k][k];
      matrix[k][k] = 1.0;
 
      for (int i=0; i<2; i++) {
@@ -31,7 +31,7 @@ mat_invert2x2(double matrix[][2])
 }
 
 int
-mat_mul2x2(double A[][2], double B[][2], double C[][2])
+mat_mul2x2(float A[][2], float B[][2], float C[][2])
 {
   for (int i=0; i<2; i++)
     for (int j=0; j<2; j++){
@@ -44,7 +44,7 @@ mat_mul2x2(double A[][2], double B[][2], double C[][2])
 }
 
 int
-mat_mul2x1(double A[][2], double *B, double *C)
+mat_mul2x1(float A[][2], float *B, float *C)
 {
   C[0] = A[0][0]*B[0] + A[0][1]*B[1];
   C[1] = A[1][0]*B[0] + A[1][1]*B[1];
@@ -53,7 +53,7 @@ mat_mul2x1(double A[][2], double *B, double *C)
 }
 
 int
-mat_add2x2(double A[][2], double B[][2], double C[][2])
+mat_add2x2(float A[][2], float B[][2], float C[][2])
 {
   for (int i=0; i<2; i++){
     for (int j=0; j<2; j++){
@@ -65,7 +65,7 @@ mat_add2x2(double A[][2], double B[][2], double C[][2])
 }
 
 int
-vec_add2x1(double *A, double *B, double *C)
+vec_add2x1(float *A, float *B, float *C)
 {
   for (int i=0; i<2; i++){
     C[i] = A[i] + B[i];
@@ -75,7 +75,7 @@ vec_add2x1(double *A, double *B, double *C)
 }
 
 int
-vec_sub2x1(double *A, double *B, double *C)
+vec_sub2x1(float *A, float *B, float *C)
 {
   for (int i=0; i<2; i++){
     C[i] = A[i] - B[i];
@@ -85,7 +85,7 @@ vec_sub2x1(double *A, double *B, double *C)
 }
 
 int
-mat_sub2x2(double A[][2], double B[][2], double C[][2])
+mat_sub2x2(float A[][2], float B[][2], float C[][2])
 {
   for (int i=0; i<2; i++){
     for (int j=0; j<2; j++){
@@ -97,7 +97,7 @@ mat_sub2x2(double A[][2], double B[][2], double C[][2])
 }
 
 int
-mat_copy2x2(double A[][2], double B[][2])
+mat_copy2x2(float A[][2], float B[][2])
 {
   for (int i=0; i<2; i++){
     for (int j=0; j<2; j++){
@@ -109,7 +109,7 @@ mat_copy2x2(double A[][2], double B[][2])
 }
 
 int
-mat_iden2x2(double A[][2])
+mat_iden2x2(float A[][2])
 {
   for (int i=0; i<2; i++){
     for (int j=0; j<2; j++){
